add filewatch msvc project

+++ clibs/filewatch/luafilewatch.cpp
@@ -43,12 +43,12 @@ static std::wstring u2w(const strview& str) {
     if (str.empty()) {
         return L"";
     }
-    int wlen = ::MultiByteToWideChar(CP_UTF8, 0, str.data(), str.size(), NULL, 0);
+    int wlen = ::MultiByteToWideChar(CP_UTF8, 0, str.data(), (int)str.size(), NULL, 0);
     if (wlen <= 0) {
         return L"";
     }
     std::vector<wchar_t> result(wlen);
-    ::MultiByteToWideChar(CP_UTF8, 0, str.data(), str.size(), result.data(), wlen);
+    ::MultiByteToWideChar(CP_UTF8, 0, str.data(), (int)str.size(), result.data(), wlen);
     return std::wstring(result.data(), result.size());
 }
 
@@ -56,12 +56,12 @@ static std::string w2u(const wstrview& wstr) {
     if (wstr.empty()) {
         return "";
     }
-    int len = ::WideCharToMultiByte(CP_UTF8, 0, wstr.data(), wstr.size(), NULL, 0, 0, 0);
+    int len = ::WideCharToMultiByte(CP_UTF8, 0, wstr.data(), (int)wstr.size(), NULL, 0, 0, 0);
     if (len <= 0) {
         return "";
     }
     std::vector<char> result(len);
-    ::WideCharToMultiByte(CP_UTF8, 0, wstr.data(), wstr.size(), result.data(), len, 0, 0);
+    ::WideCharToMultiByte(CP_UTF8, 0, wstr.data(), (int)wstr.size(), result.data(), len, 0, 0);
     return std::string(result.data(), result.size());
 }
